Reorganize the validates module
