+++ creusot/src/translation/function.rs
@@ -339,7 +339,7 @@ impl<'body, 'tcx> BodyTranslator<'body, 'tcx> {
         );
     }
 
-    fn emit_ghost_assign(&mut self, lhs: Place<'tcx>, rhs: Term<'tcx>, span: Span) {
+    fn emit_snapshot_assign(&mut self, lhs: Place<'tcx>, rhs: Term<'tcx>, span: Span) {
         self.emit_assignment(&lhs, fmir::RValue::Ghost(rhs), span)
     }
